Bugfix commit for Linux and Mac builds.
This commit just fixes minor bugs to get Linux and Mac builds to work with GCC and Apple Clang. You have to use the most up-to-date version of Apple Clang (at least Clang version 14.0.3) because otherwise it does not instantiate constexpr functions correctly even though it should work per the language standard.

+++ usersettings.cpp
@@ -283,7 +283,7 @@ struct UserSettingsImpl : public UserSettings
         sets.setValue("defaultalgo", int(defaultAlgo));
         sets.setValue("entersubdirs", navigateSubs);
         sets.setValue("displayuppercase", useUppercase);
-        sets.setValue("maxfiles", maxFiles);
+        sets.setValue("maxfiles", qulonglong(maxFiles));
     }
 
 	void clearAllSettings() override
